Restore original MemGen and reorganize experiment pipeline
- Add .gitignore for backup folders and outputs
- Reorganize gsm8k_pipeline scripts by model (Qwen3-8B, SmolLM3-3B)
- Add experiment configs and run_pipeline.py
- Restore original MemGen implementation (remove _grpo_forward, is_grpo flag)
- Fix projection layer and query_latents dtype issues
- Restore chat_template and mixed_precision settings

+++ memgen/model/weaver.py
@@ -21,13 +21,13 @@ class MemGenWeaver(nn.Module):
 
         # prompt augmentation
         self.prompt_query_latents = nn.Parameter(
-            torch.randn(prompt_latents_len, self.hidden_size, dtype=torch.bfloat16),
+            torch.randn(prompt_latents_len, self.hidden_size),
             requires_grad=True
         )
 
         # inference augmentation
         self.inference_query_latents = nn.Parameter(
-            torch.randn(inference_latents_len, self.hidden_size, dtype=torch.bfloat16),
+            torch.randn(inference_latents_len, self.hidden_size),
             requires_grad=True
         )
